Move `VState` into a separate module
