+++ eras/conway/impl/src/Cardano/Ledger/Conway/State.hs
@@ -2,9 +2,11 @@ module Cardano.Ledger.Conway.State (
   module Cardano.Ledger.Babbage.State,
   module Cardano.Ledger.Conway.State.CertState,
   module Cardano.Ledger.Conway.State.Stake,
+  module Cardano.Ledger.Conway.State.VState,
 ) where
 
 import Cardano.Ledger.Babbage.State
 
 import Cardano.Ledger.Conway.State.CertState
 import Cardano.Ledger.Conway.State.Stake
+import Cardano.Ledger.Conway.State.VState
